Now analyzes potential chunks
Program is in an in-between state, still rendering from a finite grid but preparing for infinite world generation. Right now it takes the given grid and looks at potential chunks it could group blocks into. The chunks use the same (quasiperiodic) grid so that it's all nestable. Need to figure out whether the 30 chunk types it's currently generating will end up working for me.